Move register route to dedicated file
